stopped data race
