+++ src/config/system_stats/mod.rs
@@ -2,5 +2,6 @@ pub mod battery;
 pub mod cores;
 pub mod cpu;
 pub mod disk;
-pub mod network;
+pub mod network_down;
+pub mod network_up;
 pub mod ram;
